añadir los  productos al carro de compras

+++ src/main/java/com/ecommerce/model/Producto.java
@@ -39,9 +39,12 @@ public class Producto {
     private String descripcion;
     private String imagen;
     private double precio;
-    private int cantidad;
+    private double cantidad;
     
     @ManyToOne
     private Usuario usuario;
+    
+    
+    
 
 }
